refactor: move SimpleSlider to ui/ directory
