+++ src/components/ui/SimpleSlider.vue
