+++ src/views/HomeView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import SimpleSlider from '@/components/features/SimpleSlider.vue'
+import SimpleSlider from '@/components/ui/SimpleSlider.vue'
 import ProductList from '@/components/features/ProductList.vue'
 </script>
